# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Chat de Atendimento - API
  description: |
    API RESTful para sistema de atendimento ao cliente via WhatsApp.
    
    ## Funcionalidades Principais
    
    - 🔐 **Autenticação JWT** com SQL Server
    - 💬 **WhatsApp Business API** (Meta Cloud)
    - 🤖 **Chatbot IA** com Google Gemini
    - ⚡ **WebSocket** para tempo real
    - 📊 **Dashboard** com métricas
    
    ## Fluxo de Autenticação
    
    1. Faça login via `POST /auth/login`
    2. Use o token retornado no header `Authorization: Bearer <token>`
    3. Token expira em 24 horas
    
  version: 2.0.0
  contact:
    name: Suporte API
    email: api@empresa.com
    url: https://empresa.com/suporte
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Ambiente de Desenvolvimento
  - url: https://api.empresa.com/api/v1
    description: Ambiente de Produção
...
paths:
...
  /users:
    get:
      tags:
        - users
      summary: Listar usuários
      description: Lista todos os usuários (somente admin)
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Número da página
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Itens por página
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: role
          in: query
          description: Filtrar por role
          schema:
            type: string
            enum: [admin, atendente, supervisor, user]
        - name: is_active
          in: query
          description: Filtrar por status
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
        '403':
          description: Sem permissão

    post:
      tags:
        - users
      summary: Criar usuário
      description: Cria um novo usuário (somente admin)
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Dados inválidos
        '403':
          description: Sem permissão
...
  /conversations:
    get:
      tags:
        - conversations
      summary: Listar conversas
      description: Lista conversas do atendente logado
      operationId: listConversations
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [queue, in_progress, resolved, closed]
        - name: channel
          in: query
          schema:
            type: string
            enum: [whatsapp, web, api]
        - name: assigned_to
          in: query
          schema:
            type: integer
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Lista de conversas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationList'
...
  /conversations/{conversation_id}/messages:
    post:
      tags:
        - conversations
      summary: Enviar mensagem
      description: Envia uma mensagem em uma conversa
      operationId: sendMessage
      security:
        - bearerAuth: []
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: Mensagem enviada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido via /auth/login
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: usuario@email.com
        nome:
          type: string
          example: João Silva
        role:
          type: string
          enum: [admin, atendente, supervisor, user]
          example: atendente
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
    UserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total:
          type: integer
        page:
          type: integer
        pages:
          type: integer
    CreateUserRequest:
      type: object
      required:
        - email
        - password
        - nome
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        nome:
          type: string
        role:
          type: string
          enum: [admin, atendente, supervisor, user]
          default: user
...
    Conversation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customer_phone:
          type: string
          example: "+5511999999999"
        customer_name:
          type: string
        status:
          type: string
          enum: [queue, in_progress, resolved, closed]
        channel:
          type: string
          enum: [whatsapp, web, api]
        assigned_to:
          type: integer
        started_at:
          type: string
          format: date-time
        last_message_at:
          type: string
          format: date-time
        messages_count:
          type: integer
    ConversationList:
      type: object
      properties:
        conversations:
          type: array
          items:
            $ref: '#/components/schemas/Conversation'
        total:
          type: integer
...
    Message:
      type: object
      properties:
        id:
          type: string
        direction:
          type: string
          enum: [incoming, outgoing]
        content:
          type: string
        type:
          type: string
          enum: [text, image, document, template]
        timestamp:
          type: string
          format: date-time
        status:
          type: string
          enum: [sent, delivered, read, failed]
        sent_by:
          type: string
          enum: [agent, bot, customer]
    SendMessageRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
        type:
          type: string
          enum: [text, image, document]
          default: text
...
security:
  - bearerAuth: []